feat: 실시간 수어 번역 관련 함수 훅으로 구분
